Migrate Vue.js to Node.js/Yarn. Use vue SFCs instead of app component .js files. Get rid of apiUrl parameter. Change requirements dir location. Add Django webpack-loader.
