--- apps/static/main.js
@@ -1,9 +1,0 @@
-const app = Vue.createApp({
-	data() {
-		return {
-			subject: "English",
-		};
-	},
-});
-
-app.config.compilerOptions.delimiters = ["{(", ")}"];
