+++ apps/templates/base.html
@@ -33,17 +33,10 @@
 
 	<link href="{% static 'style.css' %}" rel="stylesheet" />
 
-	<script src="https://unpkg.com/vue@3.2.23/dist/vue.global.js"></script>
-	<script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-	<script>
-		axios.defaults.xsrfCookieName = "csrftoken";
-		axios.defaults.xsrfHeaderName = "X-CSRFTOKEN";
-	</script>
+
 	<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
 		integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous">
 	</script>
-
-	<!-- <script> axios.defaults.baseURL = "{{api_url}}";</script> -->
 </head>
 
 <body>
