--- config/requirements/base.txt
@@ -1,4 +1,0 @@
-
-# https://www.django-rest-framework.org/
-
-django- rest-framework
