+++ requirements/base.txt
@@ -1,0 +1,8 @@
+# https://docs.djangoproject.com/en/stable/
+Django==4.0.1
+
+# https://www.django-rest-framework.org/tutorial/quickstart/
+djangorestframework==3.13.1
+
+# https://github.com/django-webpack/django-webpack-loader
+django-webpack-loader==1.4.1
