+++ vue/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "vue",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build"
+  },
+  "dependencies": {
+    "@vue/cli": "^4.5.15",
+    "axios": "^0.25.0",
+    "vue": "^3.0.0",
+    "webpack": "^5.68.0",
+    "webpack-bundle-tracker": "^1.4.0"
+  },
+  "devDependencies": {
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0"
+  }
+}
