+++ vue/src/pages/subject/SubjectApp.vue
@@ -1,0 +1,13 @@
+<template>
+  <Subject />
+</template>
+
+<script>
+import Subject from "../../components/Subject.vue";
+
+export default {
+  components: {
+    Subject,
+  },
+};
+</script>
